Рефакторит и перемещает функцию getDuration в common.js
