+++ src/components/event.js
@@ -1,4 +1,4 @@
-import {getFormatTime, getFormatDate} from "../utils/common.js";
+import {getFormatTime, getFormatDate, getDuration} from "../utils/common.js";
 import AbstractComponent from "./abstract-component.js";
 
 const generateOptionsElement = (title, price) => {
@@ -20,20 +20,6 @@ const generateOptionsMarkup = (options) => {
   return ``;
 };
 
-const getDuration = (start, end) => {
-  const duration = (end.getTime() - start.getTime()) / (1000 * 60);
-  const days = Math.floor(duration / (60 * 24));
-  const hours = Math.floor((duration % (60 * 24)) / 60);
-  const minutes = Math.floor(((duration % (60 * 24)) % 60));
-
-  if (days > 0) {
-    return days + `D ` + hours + `H ` + minutes + `M`;
-  } else if (hours >= 1) {
-    return hours + `H ` + minutes + `M`;
-  }
-  return minutes + `M`;
-};
-
 const createEventTemplate = (event) => {
   const {type, city, start, end, price, options} = event;
 
@@ -43,7 +29,7 @@ const createEventTemplate = (event) => {
   const startTime = getFormatTime(start);
   const endDateTime = getFormatDate(end, `-`);
   const endTime = getFormatTime(end);
-  const duration = getDuration(start, end);
+  const duration = getDuration(event);
 
   return (
     `<li class="trip-events__item">
