+++ src/utils/common.js
@@ -25,5 +25,21 @@ const getMarkupFromArray = (array, elementCreator) => {
     .join(`\n`);
 };
 
+const getDuration = (event) => {
+  const start = event.start;
+  const end = event.end;
+  const duration = (end.getTime() - start.getTime()) / (1000 * 60);
+  const days = Math.floor(duration / (60 * 24));
+  const hours = Math.floor((duration % (60 * 24)) / 60);
+  const minutes = Math.floor(((duration % (60 * 24)) % 60));
 
-export {castTimeFormat, getFormatTime, getFormatDate, getRandomBoolean, getMarkupFromArray};
+  if (days > 0) {
+    return days + `D ` + hours + `H ` + minutes + `M`;
+  } else if (hours >= 1) {
+    return hours + `H ` + minutes + `M`;
+  }
+  return minutes + `M`;
+};
+
+
+export {castTimeFormat, getFormatTime, getFormatDate, getRandomBoolean, getMarkupFromArray, getDuration};
